nexor-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.4
slowapi==0.1.9
python-dotenv==1.0.1
python-jose==3.3.0
eth-account==0.11.1
web3==6.16.0
apscheduler==3.10.4
httpx==0.27.0
openai==1.23.6
